# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentHub Multi-Agent Chatbot Framework",
    "description": "Production-ready multi-tenant chatbot framework using LangChain 0.3+",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/tenants/{tenant_id}/permissions": {
      "get": {
        "tags": [
          "admin-tenants",
          "admin-tenants"
        ],
        "summary": "Get Tenant Permissions",
        "description": "Get all permissions (enabled agents and tools) for a tenant.\n\nRequires admin role in JWT.",
        "operationId": "get_tenant_permissions_api_admin_tenants__tenant_id__permissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantPermissionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-tenants",
          "admin-tenants"
        ],
        "summary": "Update Tenant Permissions",
        "description": "Update tenant permissions (enable/disable agents and tools for a tenant).\n\nThis will create permission records if they don't exist, or update if they do.\nRequires admin role in JWT.",
        "operationId": "update_tenant_permissions_api_admin_tenants__tenant_id__permissions_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Generic message response."
      },
      "PermissionUpdateRequest": {
        "properties": {
          "agent_permissions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Permissions",
            "description": "List of {agent_id: str, enabled: bool}"
          },
          "tool_permissions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Permissions",
            "description": "List of {tool_id: str, enabled: bool}"
          }
        },
        "type": "object",
        "title": "PermissionUpdateRequest",
        "description": "Request to update tenant permissions."
      },
...
      "TenantPermissionsResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "enabled_agents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Enabled Agents"
          },
          "enabled_tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Enabled Tools"
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "TenantPermissionsResponse",
        "description": "Response for tenant permissions."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}